Refactored conftest.py
